+++ examples/ibm-satellite/modules/route/versions.tf
@@ -14,7 +14,7 @@ provider "restapi" {
   uri   = fileexists("token.log") ? var.cluster_master_url : "test.com"
   debug = true
   headers = {
-    Authorization = format("Bearer %v",chomp(element(tolist(data.local_file.token_file.*.content), 0)))
+    Authorization = var.is_endpoint_provision ? format("Bearer %v",chomp(element(tolist(data.local_file.token_file.*.content), 0))) : ""
   }
 }