+++ ibm/data_source_ibm_is_virtual_endpoint_gateways.go
@@ -46,6 +46,11 @@ func dataSourceIBMISEndpointGateways() *schema.Resource {
 							Computed:    true,
 							Description: "The resource group id",
 						},
+						isVirtualEndpointGatewayCRN: {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "The CRN for this Endpoint Gateway",
+						},
 						isVirtualEndpointGatewayCreatedAt: {
 							Type:        schema.TypeString,
 							Computed:    true,
@@ -95,6 +100,11 @@ func dataSourceIBMISEndpointGateways() *schema.Resource {
 										Type:        schema.TypeString,
 										Computed:    true,
 										Description: "The target name",
+									},
+									isVirtualEndpointGatewayTargetCRN: {
+										Type:        schema.TypeString,
+										Computed:    true,
+										Description: "The target crn",
 									},
 									isVirtualEndpointGatewayTargetResourceType: {
 										Type:        schema.TypeString,
@@ -149,6 +159,7 @@ func dataSourceIBMISEndpointGatewaysRead(d *schema.ResourceData, meta interface{
 		endpointGatewayOutput[isVirtualEndpointGatewayHealthState] = *endpointGateway.HealthState
 		endpointGatewayOutput[isVirtualEndpointGatewayLifecycleState] = *endpointGateway.LifecycleState
 		endpointGatewayOutput[isVirtualEndpointGatewayResourceGroupID] = *endpointGateway.ResourceGroup.ID
+		endpointGatewayOutput[isVirtualEndpointGatewayCRN] = *endpointGateway.CRN
 		endpointGatewayOutput[isVirtualEndpointGatewayVpcID] = *endpointGateway.VPC.ID
 		endpointGatewayOutput[isVirtualEndpointGatewayTarget] =
 			flattenEndpointGatewayTarget(endpointGateway.Target.(*vpcv1.EndpointGatewayTarget))